Update installing.rst
Hi Sailoog,
I just tried installing OP on an Ubuntu installation. I have no experience with OP or Raspberry Pi and was just trying to follow the instructions. I found that in the Advanced section it was more or less assumed that you know how to install OP after installing the openplotter-settings app, which for me certainly wasn't the case. I had to go through a lot of forum posts to put the pieces together.
Therefore I updated the instructions page with the information I found. Hopefully it is of use to someone else.

Best regards,
Han.